third_party/WebKit/LayoutTests/fast/events/dispatch-mouse-events-to-window-always-expected.txt

Test that wheel and mouse events are dispatched to document and window even if they do not hit any element in the page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


outside body, inside element
eventType: mousedown
PASS objectsToString(actualReceivers) is "HTMLDivElement,HTMLBodyElement,HTMLDocument,Window"
eventType: mouseup
PASS objectsToString(actualReceivers) is "HTMLDivElement,HTMLBodyElement,HTMLDocument,Window"
eventType: click
PASS objectsToString(actualReceivers) is "HTMLDivElement,HTMLBodyElement,HTMLDocument,Window"
eventType: wheel
PASS objectsToString(actualReceivers) is "HTMLDivElement,HTMLBodyElement,HTMLDocument,Window"
inside body, outside element
eventType: mousedown
PASS objectsToString(actualReceivers) is "HTMLBodyElement,HTMLDocument,Window"
eventType: mouseup
PASS objectsToString(actualReceivers) is "HTMLBodyElement,HTMLDocument,Window"
eventType: click
PASS objectsToString(actualReceivers) is "HTMLBodyElement,HTMLDocument,Window"
eventType: wheel
PASS objectsToString(actualReceivers) is "HTMLBodyElement,HTMLDocument,Window"
outside body, outside element, inside frame
eventType: mousedown
PASS objectsToString(actualReceivers) is "HTMLDocument,Window"
eventType: mouseup
PASS objectsToString(actualReceivers) is "HTMLDocument,Window"
eventType: click
PASS objectsToString(actualReceivers) is "HTMLDocument,Window"
eventType: wheel
PASS objectsToString(actualReceivers) is "HTMLDocument,Window"
PASS successfullyParsed is true

TEST COMPLETE
